add: camera, ui, mod: transform(local, world)

+++ Rect2DComponent.h
@@ -4,10 +4,10 @@ class Rect2DComponent :
 	public Component
 {
 public:
-	unsigned int sx, sy;
+	float sx, sy;
 	unsigned int color;
 	
-	Rect2DComponent(unsigned int _sx, unsigned int _sy, unsigned int _color) : sx(_sx), sy(_sy), color(_color) {}
+	Rect2DComponent(float _sx, float _sy, unsigned int _color) : sx(_sx), sy(_sy), color(_color) {}
 
 	void Render(GameObject& obj) override;
 };
